feat(common): add json schema convert utility
