--- core/services/_test/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "test",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "dev": "nodemon index.js",
-    "dev2": "nodemon refResolver.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@apidevtools/json-schema-ref-parser": "^9.0.9",
-    "ajv": "^8.6.2",
-    "ajv-formats": "^2.1.1",
-    "axios": "^0.21.1",
-    "chalk": "^4.1.2",
-    "lodash": "^4.17.21",
-    "url-join": "^4.0.1"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.12"
-  }
-}
