+++ core/services/common/package.json
@@ -4,12 +4,12 @@
   "description": "",
   "main": "api/expressServer.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node api/expressServer.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@apidevtools/json-schema-ref-parser": "^9.0.9",
     "@elastic/elasticsearch": "^7.13.0",
     "ajv": "^8.6.2",
     "ajv-formats": "^2.1.0",
